+++ CHANGELOG.rst
@@ -108,6 +108,8 @@ Added
   `#198 <https://github.com/openego/eGon-data/issues/198>`_
 * Integrate data bundle
   `#272 <https://github.com/openego/eGon-data/issues/272>`_
+* Integrate power-to-gas installation potential links
+  `#293 <https://github.com/openego/eGon-data/issues/293>`_
 * Integrate distribution of wind onshore and pv ground mounted generation
   `#146 <https://github.com/openego/eGon-data/issues/146>`_
 * Integrate dynamic line rating potentials
@@ -116,7 +118,12 @@ Added
   `#308 <https://github.com/openego/eGon-data/issues/308>`_
 * Integrate supply strategies for individual and district heating
   `#232 <https://github.com/openego/eGon-data/issues/232>`_
-
+* Integrate gas production
+  `#321 <https://github.com/openego/eGon-data/issues/321>`_
+* Integrate industrial time series creation
+  `#237 <https://github.com/openego/eGon-data/issues/237>`_
+* Merge electrical loads per bus and export to etrago tables
+  `#328 <https://github.com/openego/eGon-data/issues/328>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159
 
@@ -160,8 +167,24 @@ Changed
   `#281 <https://github.com/openego/eGon-data/issues/281>`_
 * Migrate mastr, mv_grid_districts and re_potential_areas to datasets
   `#297 <https://github.com/openego/eGon-data/issues/297>`_
-* Rename etrago tables from e.g. egon_pf_hv_bus to egon_etrago bus etc. 
+* Migrate industrial sites to datasets
+  `#237 <https://github.com/openego/eGon-data/issues/237>`_
+* Rename etrago tables from e.g. egon_pf_hv_bus to egon_etrago bus etc.
   `#334 <https://github.com/openego/eGon-data/issues/334>`_
+* Move functions used by multiple datasets
+  `#323 <https://github.com/openego/eGon-data/issues/323>`_
+* Migrate scenario tables to datasets
+  `#309 <https://github.com/openego/eGon-data/issues/309>`_
+* Migrate weather data and power plants to datasets
+  `#314 <https://github.com/openego/eGon-data/issues/314>`_
+* Create and fill table for CTS electricity demand per bus
+  `#326 <https://github.com/openego/eGon-data/issues/326>`_
+* Migrate osmTGmod to datasets
+  `#305 <https://github.com/openego/eGon-data/issues/305>`_
+* Filter osm landuse areas, rename industrial sites tables and update load curve function
+  `#378 <https://github.com/openego/eGon-data/issues/378>`_
+* Remove version columns from eTraGo tables and related code
+  `#384 <https://github.com/openego/eGon-data/issues/384>`_
 
 Bug fixes
 ---------
@@ -198,3 +221,8 @@ Bug fixes
   `#340 <https://github.com/openego/eGon-data/issues/340>`_
 * Set current working directory as java's temp dir when executing osmosis
   `#344 <https://github.com/openego/eGon-data/issues/344>`_
+* Fix border gas voronoi polygons which had no bus_id
+  `#362 <https://github.com/openego/eGon-data/issues/362>`_
+* Add dependency from WeatherData to Vg250
+  `#387 <https://github.com/openego/eGon-data/issues/387>`_
+